Fixed 2D table axiis not incrementing
git-svn-id: http://svn.3splooges.com/romraider-arch/trunk@141 d2e2e1cd-ba16-0410-be16-b7c4453c7c2d

+++ src/enginuity/maps/Table2D.java
@@ -91,7 +91,7 @@ public class Table2D extends Table {
         //this.colorize();     
     }
     
-    public void increment(int increment) {
+    public void increment(double increment) {
         super.increment(increment);
         axis.increment(increment);
     }
